[fix] 1、动态库转静态库 2、全都静态链接运行库 3、修复所有warning

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -2,8 +2,6 @@
 #include "ImGuiLayer.h"
 
 #include "imgui.h"
-
-
 #include "backends/imgui_impl_glfw.h"
 #include "backends/imgui_impl_opengl3.h"
 
@@ -73,7 +71,7 @@ namespace Hazel {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		// Rendering
 		ImGui::Render();
